Use a delegate handler for DiffFrame listening to a.click
Refactor krRTRC_RebindElements to no longer have to bind a
click handler each time the frame is opened.

Instead bind a delegate handler when the frame is first created
so that it automatically gets all link clicks inside (incl. for
links created inside the frame in the future).

What remains in krRTRC_RebindElements is applying classes for
rtrc items. Those should be done at update time right before
inserting the new feed in the document instead of bruteforced
afterwards.