Don't hide the toolbox portlet
This used to be hidden to make the sidebar smaller,
but since we now allow toggling of the sidebar as a whole
there is no longer a benefit in hiding this unconditionally.

In fact, it made it rather confusing when expanding the sidebar
as the Toolbox would still be missing from view.