diff: Fire 'wikipage.diff' hook on each loaded diff
Only applied to diff frame content for edits. While new pages
also use the 'DiffFrame' element to display the newly created page,
it is not a diff.

Also, don't run it when MassPatrol is enabled since the diff is
expected to be short-lived in that case (in the future, we may
want to make MassPatrol patrol the edit without rendering the
diff first, e.g. just marking it in the list).

Fixes #9