settings: Make 'Edits' and 'New pages' on by default
The default behaviour is to show both edits and new pages. It's just
that the checkboxes and internal values were false by default and
it was relying on the default of getApiRcParams() to fetch 'edits|new'
in that case. However that was meant to cover in the event the user
unticked both, not the common case.

This way better reflects what is happening to the user.