feed: Cancel pending request when applying new settings
When applying new settings (or pause/unpause), try to cancel any request that
is already initiated from automatic refresh interval.

This is also in preparation for enabling kickstart behaviour by default, which
increases the likelihood of users applying settings shortly after a request has
started. If the user is on a relatively slow network, this would improve
performance by not needlessly waiting for the now-irrelevant request to finish.

Follows-up 45f1e87. It must wakeup MassPatrol after the first request with the
new settings is finished. It should not have reused the promise for an existing
request that started before the new settings were applied. The reuse made it
resolve faster but with incorrect behaviour.