layout: Move navtoggle space to mw-panel to avoid jiggling bug
Follows-up 2a2b360 which didn't account for the fact that the navtoggle is
mirrored by scaleX(-1) on hover, which makes this spacer go to the other
side. This could be fixed by making the spacer much larger, or change size but
that adds other complexity and might introduce other bugs.