settings: Fix inheritance of opt>defOpt.
Follows-up 3afc181868ac392, which used Object.create() for the
parent object (defOpt) instead of the child/leaf objects (rc, app).

As such, when the form elements were changed by a user, it was
actually being read into the 'defOpt' object instead of the 'opt'
object.

The impact of that is that the permalink, which is based on
a diff between 'opt' and 'defOpt' was always empty.

Fix by using $.extend(true, ..) as before. The change to
create() was a micro- optimisation with no tangible benefit.

Fixes #80.