diff: Hide broken "Browse history interactively" link
This is inserted server-side by the RevisionSlider extension, but
doesn't work. And even if it did, it is not what we want to expose
in the context of reviewing a specific edit in RTRC.

(The anchor tag has no "href" attribute, clicking it is a no-op.)

Fixes https://github.com/Krinkle/mw-gadget-rtrc/issues/98.