diff: Restore "new window" behaviour for core's tool links
Previously we relied on the fact that core's tool links are absolute
URLs to distinguish them from our own interactive links, and links
by other gadgets and such.

As result of https://phabricator.wikimedia.org/T263581, this no longer
works that way. It's possible this will be fixed upstream, but for
RTRC we can make it work in a different way by moving the click handler
to be registered last, and thus apply after our other handles that
stop event propagation, and thus will ensure this last handler is
not going to be called.

Fixes https://github.com/Krinkle/mw-gadget-rtrc/issues/99
Ref https://nl.wikipedia.org/w/index.php?oldid=60037002#RTRC